internal: Pull examples into workspace for CI
