--- packages/core/src/state/RIC.d.ts
@@ -1,2 +1,0 @@
-declare const RIC: (cb: (...args: any[]) => void, options: any) => void;
-export default RIC;
