--- packages/core/src/state/RIC.js
@@ -1,5 +1,0 @@
-const RIC =
-  typeof requestIdleCallback === 'function'
-    ? requestIdleCallback
-    : cb => setTimeout(cb, 0);
-export default RIC;
