+++ packages/core/src/state/RIC.ts
@@ -1,0 +1,6 @@
+const RIC: (cb: (...args: any[]) => void, options: any) => void = (
+  typeof requestIdleCallback === 'function'
+    ? requestIdleCallback
+    : (cb: any) => setTimeout(cb, 0)
+) as any;
+export default RIC;
