+++ packages/endpoint/tsconfig.json
@@ -3,7 +3,6 @@
   "compilerOptions": {
     "outDir": "lib",
     "rootDir": "src",
-    "skipLibCheck": false
   },
   "include": ["src", "typescript-tests"],
   "references": [{ "path": "../../__tests__" }]
