+++ packages/normalizr/tsconfig.json
@@ -4,7 +4,6 @@
     "outDir": "lib",
     "rootDir": "src",
     "noImplicitAny": false,
-    "skipLibCheck": false
   },
   "include": ["src"]
 }
