+++ tsconfig-base.json
@@ -41,8 +41,9 @@
     "moduleResolution": "node",            /* Specify module resolution strategy: 'node' (Node.js) or 'classic' (TypeScript pre-1.6). */
     "moduleSuffixes": ["", ".native"],
     // "rootDirs": ["./src"],                        /* List of root folders whose combined content represents the structure of the project at runtime. */
-    // "typeRoots": [],                       /* List of folders to include type definitions from. */
-     //"types": [],                           /* Type declaration files to be included in compilation. */
+    //"typeRoots": [],                       /* List of folders to include type definitions from. */
+    // explicitly only include these @types/ libs
+    //"types": ["node","jest","react","react-dom"],                           /* Type declaration files to be included in compilation. */
     "allowSyntheticDefaultImports": true,     /* Allow default imports from modules with no default export. This does not affect code emit, just typechecking. */
     "esModuleInterop": true,                   /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */
     // "preserveSymlinks": true,              /* Do not resolve the real path of symlinks. */
@@ -53,6 +54,7 @@
     "paths": {
       "__tests__/*": ["__tests__/*"]
     },
+    "skipDefaultLibCheck": true,
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true,
